#42 little refactoring
