+++ NodeKit/Chains/UrlServiceChainBuilder.swift
@@ -24,14 +24,14 @@ open class UrlServiceChainBuilder {
     }
 
     /// Создает цепочку узлов, описывающих транспортный слой обработки.
-    open func requestBsonTrasportChain(providers: [MetadataProvider], session: URLSession?) -> TransportBsonLayerNode {
+    open func requestBsonTrasportChain(providers: [MetadataProvider], session: URLSession?) -> Node<TransportUrlRequest, Bson> {
         let requestSenderNode = RequestSenderNode(rawResponseProcessor: self.urlResponseBsonProcessingLayerChain(), manager: session)
         let technicalErrorMapperNode = TechnicaErrorMapperNode(next: requestSenderNode)
         return RequestCreatorNode(next: technicalErrorMapperNode, providers: providers)
     }
 
     /// Создает цепочку узлов, описывающих транспортный слой обработки.
-    open func requestTrasportChain(providers: [MetadataProvider], session: URLSession?) -> TransportLayerNode {
+    open func requestTrasportChain(providers: [MetadataProvider], session: URLSession?) -> Node<TransportUrlRequest, Json> {
         let requestSenderNode = RequestSenderNode(rawResponseProcessor: self.urlResponseProcessingLayerChain(), manager: session)
         let technicalErrorMapperNode = TechnicaErrorMapperNode(next: requestSenderNode)
         return RequestCreatorNode(next: technicalErrorMapperNode, providers: providers)
