+++ NodeKit/Layers/LayerTypes.swift
@@ -6,10 +6,6 @@
 //  Copyright © 2019 Кравченков Александр. All rights reserved.
 //
 
-/// Явный тип для слоя транспорта с `JSON`
-public typealias TransportLayerNode = Node<TransportUrlRequest, Json>
-/// Явный тип для слоя транспорта с `BSON`
-public typealias TransportBsonLayerNode = Node<TransportUrlRequest, Bson>
 /// Явный тип для слоя обработки запроса.
 public typealias RequestProcessingLayerNode = Node<URLRequest, Json>
 /// Явный тип для слоя обработки ответа `JSON`
