+++ NodeKit/Layers/RequestBuildingLayer/Models/RequestEncodingModel.swift
@@ -1,0 +1,32 @@
+//
+//  RequestEncodingModel.swift
+//  NodeKit
+//
+//  Created by Vladislav Krupenko on 13.04.2020.
+//  Copyright © 2020 Кравченков Александр. All rights reserved.
+//
+
+import Foundation
+
+public struct RequestEncodingModel<Raw> {
+
+    /// Параметры для формирования запроса.
+    public let urlParameters: TransportUrlParameters
+    /// Данные в виде `BSON` или `JSON`
+    public let raw: Raw
+    /// Кодировка данных запроса
+    public let encoding: ParametersEncoding
+
+    /// Инцииаллизирует объект.
+    /// - Parameter urlParameters: Параметры для формирования запроса.
+    /// - Parameter raw: Данные в виде `BSON` или `JSON`
+    /// - Parameter encoding: Кодировка данных запроса
+    public init(urlParameters: TransportUrlParameters,
+                raw: Raw,
+                encoding: ParametersEncoding) {
+        self.urlParameters = urlParameters
+        self.raw = raw
+        self.encoding = encoding
+    }
+
+}
